+++ src/error.rs
@@ -1,0 +1,9 @@
+use std::{io, path::PathBuf};
+
+#[derive(Debug)]
+pub enum ResolverError {
+    Io(io::Error),
+    UnexpectedJson((PathBuf, serde_json::Error)),
+    UnexpectedValue(String),
+    ResolveFailedTag,
+}