+++ src/plugin/main_file.rs
@@ -1,17 +1,17 @@
 use std::sync::Arc;
 
-use crate::{description::PkgFileInfo, Resolver};
+use crate::{description::PkgInfo, Resolver};
 
 use super::{
     AliasFieldPlugin, ExportsFieldPlugin, ImportsFieldPlugin, Plugin, ResolveInfo, ResolverStats,
 };
 
 pub struct MainFilePlugin<'a> {
-    pkg_info: &'a Option<Arc<PkgFileInfo>>,
+    pkg_info: &'a Option<Arc<PkgInfo>>,
 }
 
 impl<'a> MainFilePlugin<'a> {
-    pub fn new(pkg_info: &'a Option<Arc<PkgFileInfo>>) -> Self {
+    pub fn new(pkg_info: &'a Option<Arc<PkgInfo>>) -> Self {
         Self { pkg_info }
     }
 }