+++ src/utils.rs
@@ -1,22 +1,8 @@
 use std::path::{Path, PathBuf};
 
-use crate::{ResolveInfo, Resolver};
-
-pub static RAISE_RESOLVE_ERROR_TAG: &str = "T0";
+use crate::Resolver;
 
 impl Resolver {
-    pub(super) fn raise_tag() -> String {
-        RAISE_RESOLVE_ERROR_TAG.to_string()
-    }
-
-    pub(super) fn raise_resolve_failed_message(info: &ResolveInfo) -> String {
-        format!(
-            "Resolve '{}' failed in '{}'",
-            info.request,
-            info.path.display()
-        )
-    }
-
     pub(super) fn find_up(now_dir: &Path, file_name: &str) -> Option<PathBuf> {
         let path = now_dir.join(file_name);
         if path.is_file() {